+++ src/containers/DungeonContainer.jsx
@@ -4,8 +4,12 @@ import {react, redux} from '../helpers/helpers';
 import dungeon from '../helpers/dungeon';
 
 import DungeonView from '../views/DungeonView';
-import PlayerContainer from "./PlayerContainer";
+import RoomContainer from '../containers/RoomContainer';
+import RoomView from '../views/RoomView';
+import TileContainer from '../containers/TileContainer';
+import TileView from '../views/TileView';
 import CameraContainer from "./CameraContainer";
+import PlayerContainer from "./PlayerContainer";
 
 class DungeonContainer extends PureComponent {
 
@@ -14,16 +18,17 @@ class DungeonContainer extends PureComponent {
   constructor(props) {
     super(props);
     this.state = {
-      rooms: null
+      dungeonInstance: null
     }
   }
 
   componentDidMount() {
-    this.props.actions.GameMenu.updateLevel(7);
+    this.props.actions.GameMenu.updateLevel(8);
   }
 
   componentWillReceiveProps(nextProps) {
-    this.setDungeon(nextProps);
+    this.updateDungeon(nextProps);
+    this.updateRoomVisibility(nextProps);
   }
 
   /* local state */
@@ -34,20 +39,59 @@ class DungeonContainer extends PureComponent {
     };
   }
 
-  getDungeonRooms(props) {
-    this.setState({ ...this.state,
-      rooms: dungeon.build(props)
+  getDungeonInstance(props) {
+    const grids = this.tileGrids.reduce((result, grid) => {
+      result.push(grid.map(tiles => tiles.map((tileProps, index) => {
+        return (
+          <TileContainer key={index}>
+            <TileView {...tileProps}></TileView>
+          </TileContainer>
+        );
+      })));
+      return result;
+    }, []);
+
+    const instance = this.roomGrids.reduce((result, roomProps, index) => {
+      const tiles = grids[index];
+      //if (props.state.Entity.Player.spawns[0].position.roomId === index) {
+        result.push(
+          <RoomContainer key={index}>
+            <RoomView {...roomProps}>{tiles}</RoomView>
+          </RoomContainer>
+        );
+      //}
+      return result;
+    }, []);
+
+    return instance;
+  }
+
+  setNewDungeon(props) {
+    const {tileGrids, roomGrids} = dungeon.build(props);
+    this.tileGrids = tileGrids;
+    this.roomGrids = roomGrids;
+
+    this.refreshDungeon(props);
+  }
+
+  refreshDungeon(props) {
+    this.setState({...this.state,
+      dungeonInstance: this.getDungeonInstance(props)
     });
   }
 
   /* updates */
 
-  setDungeon(nextProps) {
-    if (
-      react.stateDidUpdate(this.props, nextProps, 'GameBoard.hasPerpective') ||
-      react.stateDidUpdate(this.props, nextProps, 'GameBoard.level')
-    ) {
-      this.getDungeonRooms(nextProps);
+  updateDungeon(nextProps) {
+    if (react.stateDidUpdate(this.props, nextProps, 'GameBoard.hasPerpective') ||
+      react.stateDidUpdate(this.props, nextProps, 'GameBoard.level')) {
+      this.setNewDungeon(nextProps);
+    }
+  }
+
+  updateRoomVisibility(nextProps) {
+    if (react.stateDidUpdate(this.props, nextProps, 'Entity.Player.spawns[0].position.roomId')) {
+      if (!isNaN(nextProps.state.Entity.Player.spawns[0].position.roomId)) this.refreshDungeon(nextProps);
     }
   }
 
@@ -57,7 +101,7 @@ class DungeonContainer extends PureComponent {
     return (
       <DungeonView {...this.getDungeonProps(this.props)}>
         <CameraContainer>
-          <div className='rooms'>{ this.state.rooms }</div>
+          <div className='rooms'>{ this.state.dungeonInstance }</div>
           <PlayerContainer/>
           {/*<MinionContainer />*/}
         </CameraContainer>
@@ -68,9 +112,9 @@ class DungeonContainer extends PureComponent {
 
 export default connect(
   state => redux.mapState(state, [
-    'GameBoard', 'Tile'
+    'GameBoard', 'Tile', 'Entity'
   ]),
   dispatch => redux.mapActions(dispatch, [
-    'GameMenu', 'Tile'
+    'GameMenu', 'Tile', 'Entity'
   ])
 )(DungeonContainer);