+++ src/containers/LevelGeneratorContainer.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import LevelGeneratorApp from '../level-generator/level-generator.js'
+
+export default props => {
+  return <LevelGeneratorApp />
+}