+++ src/level-generator/components/TileGridInfo.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+class TileGridInfo extends React.PureComponent {
+  render() {
+    const items = [];
+    const info = this.props.gridInfo();
+    for( let key in info ) {
+      if( info.hasOwnProperty(key) ) {
+        items.push(<p key={key}>{key + ": " + info[key]}</p>)
+      }
+    }
+    return <div id="tileGridInfo">{items}</div>
+  }
+}
+
+export {TileGridInfo as default};