+++ src/redux/actions/GameMenuActions.js
@@ -1,6 +1,7 @@
 import {
-  GAME_MENU_CONSTANTS, PLAYER_CONSTANTS
+  GAME_MENU_CONSTANTS
 } from '../../constants';
+import GameMenuActions from './GameMenuActions';
 
 export default {
   togglePerspective: flag => {
@@ -18,6 +19,17 @@ export default {
         type: GAME_MENU_CONSTANTS.LEVEL_SET,
         payload: newLevel
       });
+    }
+  },
+
+  reloadLevel: () => {
+    return (dispatch, getState) => {
+      const currentLevel = getState().GameBoard.level;
+      dispatch({
+        type: GAME_MENU_CONSTANTS.LEVEL_SET,
+        payload: undefined
+      });
+      return setTimeout(() => dispatch(GameMenuActions.updateLevel(currentLevel)), 100);
     }
   }
 };