+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.0.7] - 2017-08-29
+### Added
+- Added new level generator
+
+## [0.0.6] - 2017-08-22
+### Added
+- Optimized dom rendering by only rendering player's room
+
+## [0.0.5] - 2017-08-22
+### Added
+- Randomized dungeon room layout
+
 ## [0.0.4] - 2017-08-22
 ### Added
 - Keyboard support